Finish
Buffs show their correct time on the hud.
Improved efficiency regarding usage of buff slots.
Make this branch ship-ready.

+++ GameServerCore/Domain/IStackable.cs
@@ -2,9 +2,9 @@ namespace GameServerCore.Domain
 {
     public interface IStackable
     {
-        byte StackCount { get; }
+        int StackCount { get; }
         bool IncrementStackCount();
         bool DecrementStackCount();
-        void SetStacks(byte newStacks);
+        void SetStacks(int newStacks);
     }
 }
